Update & Code Cleanup

+++ src/server.py
@@ -1,35 +1,31 @@
-import datetime
-import functools
-import logging
-import os
-
 from dotenv import load_dotenv
+from os import getenv
+from datetime import datetime
+from functools import wraps
 from flask import Flask, make_response, jsonify, request, session, g
 from flask_cors import CORS
 from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy import text
-from sqlalchemy.exc import InvalidRequestError, OperationalError, SQLAlchemyError
-from werkzeug.exceptions import BadRequest, NotFound, MethodNotAllowed, HTTPException
+from sqlalchemy.exc import SQLAlchemyError
+from werkzeug.exceptions import HTTPException
 
 from src.hashed import generate_hashed, check_hashed
 from src.schema import user_schema, users_schema, login_user_schema
 
 
 load_dotenv()
-logging.basicConfig(level=logging.DEBUG)
-server_mode = os.getenv("SERVER_MODE")
 
 
 # Flask App, Database, CORS and Session Configuration
 app = Flask(__name__)
-app.config["SQLALCHEMY_DATABASE_URI"] = os.getenv("MYSQLDB_URI")
+app.config["SQLALCHEMY_DATABASE_URI"] = getenv("MYSQLDB_URI")
 app.config["SQLALCHEMY_TRACK_MODIFICATION"] = False
 app.config["SQLALCHEMY_ECHO"] = False
-app.config["SECRET_KEY"] = os.getenv("SECRET_KEY")
+app.config["SECRET_KEY"] = getenv("SECRET_KEY")
 app.config["SESSION_TYPE"] = "filesystem"
 app.config["SESSION_PERMANENT"] = False
 app.config["SESSION_USE_SIGNER"] = True
-CORS(app, resources={r"/api/*": {"origins": os.getenv("CORS_ORIGIN")}})
+CORS(app, resources={r"/api/*": {"origins": getenv("CORS_ORIGIN")}})
 db = SQLAlchemy(app)
 
 
@@ -48,7 +44,7 @@ class Users(db.Model):
     email = db.Column(db.String(100), unique=True, nullable=False)
     fullname = db.Column(db.String(150), nullable=False)
     password = db.Column(db.String(150), nullable=False)
-    date = db.Column(db.DateTime, default=datetime.datetime.now)
+    date = db.Column(db.DateTime, default=datetime.now)
 
     def __init__(self, username, email, fullname, password):
         self.username = username
@@ -61,27 +57,31 @@ class Users(db.Model):
 
 
 # Api response function
-def api_response(message, data, code):
+def api_response(message: str, data: any, code: int):
     if data is not None:
         return make_response(jsonify({"message": message, "data": data}), code)
     return make_response(jsonify({"message": message}), code)
 
 
 # Error response function
-def error_response(error, code):
+def error_response(error: str, code: int):
     return make_response(jsonify({"error": error}), code)
+
+
+def global_session_user():
+    session_keys = ["username", "email"]
+    g.user = next((session[key] for key in session_keys if key in session), None)
 
 
 @app.before_request
 def before_request():
-    session_keys = ["username", "email"]
-    g.user = next((session[key] for key in session_keys if key in session), None)
-    print(f"Existed Session User : {g.user}")
+    global_session_user()
 
 
 @app.after_request
 def after_request(response):
-    print(f"Api Request Handled Successfully! {request.url}")
+    global_session_user()
+    print(f"Existed Session User : {g.user}")
     return response
 
 
@@ -95,6 +95,9 @@ def greet():
 def register_user():
     user_data = request.get_json()
 
+    if not user_data:
+        return error_response("User Data Required!", 400)
+
     required_fields = ["username", "email", "fullname", "password"]
     missing_fields = [field for field in required_fields if field not in user_data]
 
@@ -103,7 +106,7 @@ def register_user():
         words = input_string.split(', ')    # Split the string into a list of words
         capitalized_words = [word.capitalize() for word in words]   # Capitalize each word in the list
         result_string = ', '.join(capitalized_words)    # Join the capitalized words back into a string
-        return error_response(result_string, 403)   # Returning error response for missing fields
+        return error_response(result_string, 400)   # Returning error response for missing fields
 
     username = user_data["username"]
     email = user_data["email"]
@@ -114,11 +117,13 @@ def register_user():
     existing_user = Users.query.filter((Users.username == username) | (Users.email == email)).first()
 
     if existing_user:
+        field = None
         if existing_user.username == username:
-            return error_response("Username Already Registered", 409)
-        else:
-            return error_response("Email Already Registered", 409)
-
+            field = 'Username'
+        elif existing_user.email == email:
+            field = 'Email'
+        return error_response(f"{field} Already Registered!", 409)
+    
     # Creating and registering a new user
     new_user = Users(username, email, fullname, password)
     db.session.add(new_user)
@@ -126,6 +131,13 @@ def register_user():
 
     response = user_schema.dump(new_user)
     return api_response("User Registered Successfully!", response, 201)
+
+
+def clear_session_access_cookies() -> object:
+    session_keys = ["username", "email"]
+    session_key = next((key for key in session_keys if key in session), None)
+    session_value = session.pop(session_key, None)
+    return session_value
 
 
 # Api route for login user - http://127.0.0.1:8100/api/users/login
@@ -140,111 +152,89 @@ def login_user():
     email = user_data.get("email")
 
     if not any([username, email]):
-        return error_response("Username or Email Required!", 403)
+        return error_response("Username or Email Required!", 400)
 
     exists_user = None
+    session_key = None
 
-    if username:
-        exists_user = Users.query.filter_by(username=username).first()
-    elif email:
-        exists_user = Users.query.filter_by(email=email).first()
+    fields = {'username': username, 'email': email}
+
+    for key, value in fields.items():
+        if value:
+            session_key = key
+            exists_user = Users.query.filter_by(**{key: value}).first()
+            break
 
     if not exists_user:
-        session.pop("username", None)
-        session.pop("email", None)
+        clear_session_access_cookies()
         return error_response("User Not Found!", 404)
 
     if "password" not in user_data:
-        return error_response("Password Required!", 403)
-
+        return error_response("Password Required!", 400)
+    
     password = user_data["password"]
 
     # Verifying if same user trying too login again
-    if "username" in session and username == session["username"]:
-        return api_response("User Already Logged In!", None, 302)
-
-    elif "email" in session and email == session["email"]:
-        return api_response("User Already Logged In!", None, 302)
+    if session_key in session and (username or email) == session[session_key]:
+        return api_response(f"Welcome User, You Are Already Logged In!", None, 302)
     else:
-        # Removing exist session for new user login
-        session.pop("username", None)
-        session.pop("email", None)
-    
-    # Validating user data with given user credentials
-    if username:
-        exists_user = Users.query.filter_by(username=username).first()
-    elif email:
-        exists_user = Users.query.filter_by(email=email).first()
-    else:
-        return error_response("Invalid Credentials!", 403)
-
-    if not exists_user:
-        return error_response("User Not Found!", 404)
+        clear_session_access_cookies()
 
     check_login_user = login_user_schema.dump(exists_user)
     verify_password = check_hashed(password, check_login_user["password"])
-    print(f"Password: {password}, Status: {verify_password}")
+    status = "Success" if verify_password else "Failed"
+    print(f"Password: {password}, Status: {status}")
 
     if verify_password:
-        session_key = "username" if username else "email"
         session[session_key] = check_login_user[session_key]
         response = user_schema.dump(check_login_user)
         return api_response("User Login Successfully!", response, 202)
-    return error_response("Incorrect Password!", 406)
-
-
-# Api route for check session user - http://127.0.0.1:8100/api/users/session
-@app.route("/api/users/session", methods=["GET"])
-def get_session_user():
-    if g.user:
-        session_key = next((key for key in ["username", "email"] if key in session), None)
-
-        if session_key:
-            session_user = Users.query.filter_by(**{session_key: session[session_key]}).first()
-            if session_user:
-                response = user_schema.dump(session_user)
-                return api_response("Session User Data!", response, 200)
-            else:
-                return error_response("Session User Not Found!", 404)
-    return error_response("Unauthorized Session", 401)
+    return error_response("Incorrect Password!", 403)
 
 
 # Api route for logout current session user - http://127.0.0.1:8100/api/users/logout
 @app.route("/api/users/logout", methods=["DELETE"])
 def logout_user():
-    session_keys = ["username", "email"]
-    session_key = next((key for key in session_keys if key in session), None)
-
-    if not session_key:
-        return error_response("Unauthorized Session", 401)
-
-    session_value = session.pop(session_key, None)
+    session_value = clear_session_access_cookies()
 
     if session_value is not None:
-        return api_response("User Logged Out Successfully!", None, 206)
-    else:
-        return error_response("Unauthorized Session", 401)
+        return api_response("User Logged Out Successfully!", None, 200)
+    return error_response("Unauthorized Session", 401)
 
 
 # Login required decorator function for access other user data by session user
 def login_required(func):
-    @functools.wraps(func)
+    @wraps(func)
     def secure_function(*args, **kwargs):
         if any(key in session for key in ["username", "email"]):
+            session_key = next((key for key in ["username", "email"] if key in session), None)
+            session_user = Users.query.filter_by(**{session_key: session[session_key]}).first()
+            if session_key and session_user:
+                return func(session_user, *args, **kwargs)
             return func(*args, **kwargs)
         else:
             return error_response("Unauthorized User!", 401)
     return secure_function
 
 
+# Api route for check session user - http://127.0.0.1:8100/api/users/session
+@app.route("/api/users/session", methods=["GET"])
+@login_required
+def get_session_user(session_user, *args, **kwargs):
+    if g.user and session_user:
+        response_data = user_schema.dump(session_user)
+        return api_response("Session User Data!", response_data, 200)
+    return error_response("Unauthorized User", 401)
+
+
 # Api route for fetch all users - http://127.0.0.1:8100/api/users/fetch
 @app.route("/api/users/fetch", methods=["GET"])
 @login_required
-def fetch_users():
+def fetch_users(session_user, *args, **kwargs):
     all_users = Users.query.all()
     
     if not all_users:
-        return error_response("No Users Found!", 404)
+        return error_response("Users Not Found!", 404)
 
     response = users_schema.dump(all_users)
     return api_response("Users Fetched Successfully!", response, 200)
@@ -253,8 +243,8 @@ def fetch_users():
 # Api route for fetch user by uid - http://127.0.0.1:8100/api/users/fetch/<uid>
 @app.route("/api/users/fetch/<uid>", methods=["GET"])
 @login_required
-def fetch_user(uid):
-    existed_user = db.session.get(Users, uid)
+def fetch_user(session_user, *args, **kwargs):
+    existed_user = Users.query.get({**kwargs})
 
     if not existed_user:
         return error_response("User Not Found!", 404)
@@ -266,12 +256,12 @@ def fetch_user(uid):
 # Api route for fetch user by query parameter - http://127.0.0.1:8100/api/users/fetch/user?query=parameter
 @app.route("/api/users/fetch/user", methods=["GET"])
 @login_required
-def fetch_user_by():
+def fetch_user_by(session_user, *args, **kwargs):
     user_data = request.args.to_dict()
     print("Search User :", user_data)
 
     if not user_data:
-        return error_response("Please, Query User Data!", 406)
+        return error_response("Please, Query User Data!", 400)
 
     existed_user = Users.query.filter_by(**user_data).first()
 
@@ -284,22 +274,20 @@ def fetch_user_by():
 
 # Session user verification for modification and deletion for user data 
 def verify_session(func):
-    @functools.wraps(func)
+    @wraps(func)
     def secure_function(*args, **kwargs):
         existed_user = Users.query.filter_by(**kwargs).first()
 
         if not existed_user:
             return error_response("User Not Found!", 404)
 
-        verify_user = Users.query.filter_by(**kwargs).first()
-        response = user_schema.dump(verify_user)
-
+        response = user_schema.dump(existed_user)
         session_key = next((key for key in ["username", "email"] if key in session), None)
 
         if session_key and response[session_key] == session[session_key]:
             return func(existed_user, *args, **kwargs)
         else:
-            return error_response("Invalid or Unauthorized Session User!", 403)
+            return error_response("Invalid or Unauthorized User!", 403)
     return secure_function
 
 
@@ -335,8 +323,7 @@ def update_user(existed_user, **kwargs):
         db.session.commit()
         updated_user = Users.query.filter_by(**kwargs).first()
         response = user_schema.dump(updated_user)
-        session.pop("username", None)
-        session.pop("email", None)
+        clear_session_access_cookies()
         return api_response("User Updated Successfully!", response, 202)
     else:
         return api_response("User Data Not Modified!", None, 304)
@@ -345,52 +332,28 @@ def update_user(existed_user, **kwargs):
 # Api route for delete user by uid - http://127.0.0.1:8100/api/users/delete/<uid>
 @app.route("/api/users/delete/<uid>", methods=["DELETE"])
 @verify_session
-def delete_user(existed_user):
-    db.session.delete(existed_user)
-    db.session.commit()
-    session.pop("username", None)
-    session.pop("email", None)
-    return api_response("User Deleted Successfully!", None, 200)
+def delete_user(existed_user, *args, **kwargs):
+    session_value = clear_session_access_cookies()
+
+    if session_value is not None:
+        db.session.delete(existed_user)
+        db.session.commit()
+        return api_response("User Deleted Successfully!", None, 200)
+    return error_response("Something Went Wrong!", 400)
 
 
-# ! Basic Error/Exception handling!
-
-@app.errorhandler(BadRequest)
-def handle_bad_request_error(error):
-    return error_response(f"Bad Request: {error.description}", error.code)
-
-
-@app.errorhandler(NotFound)
-def handle_not_found_error(error):
-    return error_response(f"Not Found: {error.description} {request.url}", error.code)
-
-
-@app.errorhandler(MethodNotAllowed)
-def handle_method_not_allowed_error(error):
-    return error_response(f"Method Not Allowed: {error.description}", error.code)
-
+# Basic & SQLAlchemy Error/Exception Handling!
 
 @app.errorhandler(HTTPException)
 def handle_http_exception(error):
-    return error_response(f"HTTP Exception: {error.description}", error.code)
+    return error_response(f"{error.description}", error.code)
 
 
 @app.errorhandler(Exception)
 def handle_generic_exception(error):
-    return error_response(f"Some Error Occurred: {error}", 400)
-
-# ! SQLAlchemy Error/Exception handling!
-
-@app.errorhandler(InvalidRequestError)
-def invalid_request_error(error):
-    return error_response(f"Invalid Request: The database query is malformed! {error}", 400)
-
-
-@app.errorhandler(OperationalError)
-def operational_error(error):
-    return error_response(f"Operational Error: An issue occurred with the database operation! {error}", 500)
+    return error_response(f"Error: {error}", 400)
 
 
 @app.errorhandler(SQLAlchemyError)
 def sqlalchemy_error(error):
-    return error_response(f"SQLAlchemy Error: An unexpected issue occurred while processing your request! {error}", 500)
+    return error_response(f"SQLAlchemy Error: {error}!", 500)
